+++ fuzz/fuzz_targets/wrap_optimal_fit.rs
@@ -27,19 +27,19 @@ impl Into<OptimalFit> for Penalties {
 
 #[derive(Arbitrary, Debug, Eq, PartialEq)]
 struct Word {
-    width: u16,
-    whitespace_width: u16,
-    penalty_width: u16,
+    width: u32,
+    whitespace_width: u32,
+    penalty_width: u32,
 }
 
 #[rustfmt::skip]
 impl core::Fragment for Word {
-    fn width(&self) -> u16 { self.width }
-    fn whitespace_width(&self) -> u16 { self.whitespace_width }
-    fn penalty_width(&self) -> u16 { self.penalty_width }
+    fn width(&self) -> u32 { self.width }
+    fn whitespace_width(&self) -> u32 { self.whitespace_width }
+    fn penalty_width(&self) -> u32 { self.penalty_width }
 }
 
-fuzz_target!(|input: (u16, Vec<Word>, Penalties)| {
+fuzz_target!(|input: (u32, Vec<Word>, Penalties)| {
     let width = input.0;
     let words = input.1;
     let penalties = input.2.into();
